+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Fixed
 - Fix the `XS` size of the `dx-dialog` component
+- Fix the `disabled` state for the `dx-header` component
 
 ### Changed
 - Check and rework all storybook stories
@@ -23,6 +24,8 @@
 - Remove unused `color` property from the `dx-header` component
 - Change `dx-data-grid-generic` interface propertie types (isLoading, hasMiddlewareError, hasContentSourceAvailable, checkboxSelection)
 - Remove `dx-search-center-layout` component
+- Remove the `ignoreDisable` property from the `dx-input-textfield` component
+- Remove useless `open` property from `dx-menu` component
 
 ## 1.4.0